CC/requirements.txt
Flask==2.3.2
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0
numpy==1.23.5
requests==2.31.0
gunicorn==20.1.0
Werkzeug==2.3.3
google-cloud-storage==2.9.0
firebase-admin==6.1.0
python-dotenv==1.0.0
